== Tracker urls formatting ==
READ THIS TO AVOID A LOT OF WORK!

New way of linking to tracker pages: just use the parameter "aid" (artifact ID),
to avoid a lot of manual updates later in wiki and svn.

Example:
=========

OLD WAY TO LINK TO TRACKER
-----------------------------

Complete url of a script in Upload
http://projects.blender.org/tracker/index.php?func=detail&aid=25349&group_id=153&atid=467

If we move this in contrib this url will become
http://projects.blender.org/tracker/index.php?func=detail&aid=25349&group_id=153&atid=468
467 becomes 468, so we have to update this in wiki page.

Later on, when this moves into Trunk, the url will become
http://projects.blender.org/tracker/index.php?func=detail&aid=25349&group_id=153&atid=469
468 becomes 469, so we have to update the url in wiki page and svn.

Annoying!

NEW WAY TO LINK TO TRACKER
-----------------------------

Best way to link to tracker page is using:
http://projects.blender.org/tracker/index.php?func=detail&aid=25349

Use "func=detail"
Use "aid"            (which is the "artifact ID")

DON'T use "group_id" (which is the project ID, bf-extensions is project 153)
DON'T use "atid"     (which is the "artifacts tracker ID")

Like this, the url is unique, and we will avoid to update wiki pages and svn after moving pages
 



git-svn-id: https://svn.blender.org/svnroot/bf-extensions/trunk/py/scripts/addons@1369 96b784f8-4ae3-dd11-aff7-0030488c597c