Rigify: users can now specify layer names in the metarig armature properties.
The layer names are then used in creating the custom rig layer UI.  This is
useful for users that do not want to--or do not have the knowledge to--edit
the generated python script by hand.  It is also handy even for more advanced
users when regerating the rig over and over (which over-writes the script
and any hand-made edits).

Also misc bug fixes in some of the rig types.


git-svn-id: https://svn.blender.org/svnroot/bf-extensions/trunk/py/scripts/addons@2037 96b784f8-4ae3-dd11-aff7-0030488c597c