own patch [#30877] ply exporter to use shared faces when not exporting the normals
Faces should be exported with shared vertices whenever possible (otherwise this is simply a STL exporter). Since PLY stores normals per vertex it's understandable that we need 'individual' faces when exporting normals. However when this is not the case I think shared vertices should be default behaviour. This is what this patch implement.

Thanks Campbell for reviewing it. Thanks Aldo Zang for reporting it (off trackers, in the old fashion real life)


git-svn-id: https://svn.blender.org/svnroot/bf-extensions/trunk/py/scripts/addons@3241 96b784f8-4ae3-dd11-aff7-0030488c597c