Bunch of bug fixes.
The biggest fixes relate to keeping ID data modification out of
draw methods.  This was breaking Rigify with the current API.

Secondary fix was to move widget meshes to match the bones, even
if the widget meshes already exist.  It's nice for when the user
is progressively tweaking the metarig.


git-svn-id: https://svn.blender.org/svnroot/bf-extensions/trunk/py/scripts/addons@4277 96b784f8-4ae3-dd11-aff7-0030488c597c