Rigify: bug fix for biped arm/leg rigs.
The bbone elbow was causing weird twisting and sharp bend angles
even when smooth elbow/knee was fully disabled.  Now fixed.

This fix, however, exposes bug #34532 in Blender, so for the time
being the transition between smooth/sharp elbow is a jumpy one.
This is less objectionable, however, then the weird twisting
that was happening, as bug #34532 can be worked around by the
animator, whereas the weird twisting could not be worked around.


git-svn-id: https://svn.blender.org/svnroot/bf-extensions/trunk/py/scripts/addons@4342 96b784f8-4ae3-dd11-aff7-0030488c597c